benchmark/requirements.txt
deepeval
openai
python-dotenv
pytest
